+++ src/model-handler.cpp
@@ -21,15 +21,6 @@
 // NON-MEMBER FUNCTIONS
 // ------------------------------------------
 
-bool operator== (const Transform& a, const Transform& b)
-{
-	return
-		a.position == b.position &&
-		a.scale == b.scale &&
-		a.rotation == b.rotation &&
-		a.materialStart == b.materialStart;
-}
-
 template<typename T>
 void append_vector(std::vector<T>& origin, std::vector<T>& appendage)
 {
@@ -1046,17 +1037,6 @@ DynamicModelHashSum hash_model(const DynamicModel& model)
 	}
 	return hashSum;
 }
-
-// ------------------------------------------
-// TRANSFORM
-// ------------------------------------------
-
-Transform::Transform() :
-	position{ VECTOR_NULL }, scale{ 1, 1, 1 }, rotation{}, materialStart(0)
-{}
-Transform::Transform(Vector3 position, Vector3 scale, Quaternion rotation) :
-	position{ position }, scale{ scale }, rotation{ rotation }
-{}
 
 // ------------------------------------------
 // TINY OBJ LOADER HELPER
